+++ src/Products/PloneMeeting/migrations/migrate_to_4200.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+from collective.contact.plonegroup.utils import get_organizations
 from copy import deepcopy
 from Products.GenericSetup.tool import DEPENDENCY_STRATEGY_NEW
 from Products.PloneMeeting.migrations import logger
@@ -66,12 +67,38 @@ class Migrate_To_4200(Migrator):
            this, re-apply default values from MeetingConfigDescriptor."""
         logger.info("Configuring 'votes' for every MeetingConfigs...")
         for cfg in self.tool.objectValues('MeetingConfig'):
-            cfg.setUseVotes(MeetingConfigDescriptor.useVotes)
-            cfg.setVotesEncoder(MeetingConfigDescriptor.votesEncoder)
-            cfg.setUsedVoteValues(MeetingConfigDescriptor.usedVoteValues)
-            cfg.setFirstLinkedVoteUsedVoteValues(MeetingConfigDescriptor.firstLinkedVoteUsedVoteValues)
-            cfg.setNextLinkedVotesUsedVoteValues(MeetingConfigDescriptor.nextLinkedVotesUsedVoteValues)
-            cfg.setVoteCondition(MeetingConfigDescriptor.voteCondition)
+            config_descr = MeetingConfigDescriptor(None, None, None)
+            cfg.setUseVotes(config_descr.useVotes)
+            cfg.setVotesEncoder(config_descr.votesEncoder)
+            cfg.setUsedVoteValues(config_descr.usedVoteValues)
+            cfg.setFirstLinkedVoteUsedVoteValues(config_descr.firstLinkedVoteUsedVoteValues)
+            cfg.setNextLinkedVotesUsedVoteValues(config_descr.nextLinkedVotesUsedVoteValues)
+            cfg.setVoteCondition(config_descr.voteCondition)
+        logger.info('Done.')
+
+    def _migrateKeepAccessToItemWhenAdviceIsGiven(self):
+        """Boolean field MeetingConfig.keepAccessToItemWhenAdviceIsGiven
+           is now single select MeetingConfig.keepAccessToItemWhenAdvice."""
+        logger.info(
+            "Migrationg field 'keepAccessToItemWhenAdviceIsGiven' for "
+            "every MeetingConfigs and organizations...")
+        # MeetingConfigs
+        for cfg in self.tool.objectValues('MeetingConfig'):
+            if hasattr(cfg, 'keepAccessToItemWhenAdviceIsGiven'):
+                old_value = cfg.keepAccessToItemWhenAdviceIsGiven
+                if old_value is True:
+                    cfg.setKeepAccessToItemWhenAdvice('is_given')
+                delattr(cfg, 'keepAccessToItemWhenAdviceIsGiven')
+        # organizations
+        orgs = get_organizations(only_selected=False)
+        for org in orgs:
+            old_value = org.keep_access_to_item_when_advice_is_given
+            if old_value == '':
+                org.keep_access_to_item_when_advice = 'use_meetingconfig_value'
+            elif old_value == '1':
+                org.keep_access_to_item_when_advice = 'is_given'
+            elif old_value == '0':
+                org.keep_access_to_item_when_advice = 'default'
         logger.info('Done.')
 
     def run(self, extra_omitted=[]):
@@ -89,6 +116,7 @@ class Migrate_To_4200(Migrator):
         # configure wfAdaptations before reinstall
         self._configureItemWFValidationLevels()
         self._configureVotes()
+        self._migrateKeepAccessToItemWhenAdviceIsGiven()
 
         # reinstall so versions are correctly shown in portal_quickinstaller
         self.reinstall(profiles=['profile-Products.PloneMeeting:default', ],
@@ -119,7 +147,9 @@ def migrate(context):
     '''This migration function will:
 
        1) Configure field MeetingConfig.itemWFValidationLevels depending on old wfAdaptations;
-       2) Init otherMeetingConfigsClonableToFieldXXX new fields.
+       2) Migrate MeetingConfig.keepAccessToItemWhenAdviceIsGiven to
+          MeetingConfig.keepAccessToItemWhenAdvice;
+       3) Init otherMeetingConfigsClonableToFieldXXX new fields.
     '''
     migrator = Migrate_To_4200(context)
     migrator.run()